+++ NEWS.md
@@ -1,3 +1,7 @@
+# 1.0.1
+
+- Doc: add details about tree excluded by `qsf()`
+
 # 1.0.0
 
 Public release